Add common useForm for editing and creating goal check-ins
