+++ assets/js/features/goals/GoalCheckIn/index.tsx
@@ -2,3 +2,4 @@ export { LastCheckInMessage } from "./LastCheckInMessage";
 export { StatusSection } from "./StatusSection";
 export { DescriptionSection } from "./DescriptionSection";
 export { TargetsSection } from "./TargetsSection";
+export { useForm } from "./useForm";
